[SNT-021] New conversation이 DB에 바로 닿던 현상 변경

+++ smart_notebook_fe/src/App.tsx
@@ -45,16 +45,26 @@ function App() {
     string | null
   >(null);
   const abortControllerRef = useRef<AbortController | null>(null);
+  const newStartedThread = useRef<string | null>(null);
 
   useEffect(() => {
     loadThreads();
     loadBookmarkedThreads();
+
+    return () => {
+      newStartedThread.current = null;
+    };
   }, []);
 
   useEffect(() => {
     if (currentThread) {
-      loadMessages(currentThread.id);
-      loadBookmarks(currentThread.id);
+      if (currentThread.id !== newStartedThread.current) {
+        loadMessages(currentThread.id);
+        loadBookmarks(currentThread.id);
+      } else {
+        setMessages([]);
+        setBookmarks([]);
+      }
     }
   }, [currentThread]);
 
@@ -131,24 +141,18 @@ function App() {
 
   const createThread = async () => {
     stopMessageGeneration();
-    try {
-      const response = await fetch(`${API_BASE_URL}/threads`, {
-        method: "POST",
-        headers: getAuthHeaders(),
-        body: JSON.stringify({
-          title: "새로운 대화 / New conversation",
-        }),
-      });
-      if (response.ok) {
-        const data = await response.json();
-        const newThreads = [data.thread, ...threads].filter((t) => t != null);
-        setThreads(newThreads);
-        setCurrentThread(data.thread);
-        setMessages([]);
-      }
-    } catch (error) {
-      console.error("Error creating thread:", error);
-    }
+
+    const lastThreadId = uuidv7();
+    newStartedThread.current = lastThreadId;
+    const newThread = {
+      id: lastThreadId,
+      title: "새로운 대화 / New conversation",
+      createdAt: Date().toString(),
+      updatedAt: Date().toString(),
+    };
+
+    setThreads((prev) => [newThread, ...prev.filter((p) => p != null)]);
+    setCurrentThread(newThread);
   };
 
   const sendMessage = async (
@@ -158,8 +162,10 @@ function App() {
   ) => {
     let threadToUse = currentThread;
 
+    setMessages([...messages, { id: uuidv7(), content, role: "user" }]);
+
     // Create thread if none exists
-    if (!threadToUse) {
+    if (!threadToUse || threadToUse.id === newStartedThread.current) {
       try {
         const response = await fetch(`${API_BASE_URL}/threads`, {
           method: "POST",
@@ -171,8 +177,11 @@ function App() {
         if (response.ok) {
           const data = await response.json();
           threadToUse = data.thread;
-          setCurrentThread(threadToUse);
-          setThreads([data.thread, ...threads]);
+          if (!!threadToUse) {
+            setCurrentThread(threadToUse);
+            threads[0] = threadToUse;
+            setThreads([...threads]);
+          }
         } else {
           console.error("Failed to create thread");
           return;
@@ -189,8 +198,6 @@ function App() {
     setIsLoading(true);
 
     try {
-      setMessages([...messages, { id: uuidv7(), content, role: "user" }]);
-
       const response = await fetch(`${API_BASE_URL}/chat`, {
         method: "POST",
         headers: getAuthHeaders(),
@@ -428,6 +435,14 @@ function App() {
   };
 
   const deleteThread = async (threadId: string) => {
+    if (threadId === newStartedThread.current) {
+      setThreads((prev) => prev.filter((p) => p.id !== threadId));
+      setCurrentThread(null);
+      setMessages([]);
+      newStartedThread.current = null;
+      return;
+    }
+
     try {
       const response = await fetch(`${API_BASE_URL}/threads/${threadId}`, {
         method: "DELETE",
